% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{add_bars_to_copy_df}
\alias{add_bars_to_copy_df}
\title{Add Bars to Copy of Dataframe}
\usage{
add_bars_to_copy_df(df, metric_tibble)
}
\arguments{
\item{df}{A data frame that needs additional time marker and bar related data. The columns in the data frame can vary, but must contain a 'total_time' column. Usually received from \code{copy_and_extend_midi_df()}.}

\item{metric_tibble}{A tibble having columns 'time_steps', 'bars', and 'bar_steps'. This tibble provides the metrics that are used to add new marker and bar data to the data frame \code{df}. Created from \code{make_metric_tibble()}.}
}
\value{
A data frame which is a copy of the input data frame \code{df}, but with three additional columns: 'time_markers', 'bars', and 'bar_steps'. 'time_markers' is populated by the index of the match between the 'total_time' of \code{df} and 'time_steps' of \code{metric_tibble}. bars' and 'bar_steps' are populated from corresponding values in \code{metric_tibble} based on matching 'time_steps'.
}
\description{
This function modifies a copy of the dataframe \code{df} by adding new columns related to time markers and bars from a \code{metric_tibble}. Rows in the dataframe are iterated upon to set values for these new columns based on matching \code{time_steps} in the \code{metric_tibble}.
}
\examples{
\dontrun{
copy_track <- copy_and_extend_midi_df(midi_df, track_num = 0)
metric_tibble <- make_metric_tibble(copy_track, ticks_per_beat = 96, bars=48,smallest_tick=8)
copy_track <- add_bars_to_copy_df(copy_track,metric_tibble)
}
}
