% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-construct.R
\name{add_control_change}
\alias{add_control_change}
\title{Add control_change message to df}
\usage{
add_control_change(
  df,
  i_track = 0,
  type = "control_change",
  time = 0,
  channel = 0,
  control = 0,
  value = 0,
  ...
)
}
\arguments{
\item{df}{midi_df to add row to}

\item{i_track}{integer, default = 0}

\item{type}{character, default = "control_change"}

\item{time}{integer, default = 0}

\item{channel}{integer, default = 0}

\item{control}{integer, default = 0}

\item{value}{integer, default = 0}

\item{...}{send additional args to dplyr::add_row}
}
\value{
a new row
}
\description{
Intended for use in a dplyr style pipe. Adds a control_change message to a midi_df.
}
\examples{
\dontrun{
}
}
