% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-construct.R
\name{add_meta_end_of_track}
\alias{add_meta_end_of_track}
\title{Add end_of_track message to df}
\usage{
add_meta_end_of_track(
  meta_df,
  i_track = 0,
  type = "end_of_track",
  time = 0,
  ...
)
}
\arguments{
\item{meta_df}{midi_df to add}

\item{i_track}{integer, default = 0}

\item{type}{character, default = "end_of_track"}

\item{time}{integer, default = 0}

\item{...}{send additional args to dplyr::add_row}
}
\value{
a new row
}
\description{
Intended for use in a dplyr style pipe. Adds an end_of_track message to a meta header.
}
\examples{
\dontrun{
}
}
