% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-construct.R
\name{add_meta_tempo}
\alias{add_meta_tempo}
\title{Add set_tempo message to df}
\usage{
add_meta_tempo(
  meta_df,
  i_track = 0,
  type = "set_tempo",
  time = 0,
  tempo = 5e+05,
  ...
)
}
\arguments{
\item{meta_df}{midi_df to add}

\item{i_track}{integer, default = 0}

\item{type}{character, default = "set_tempo"}

\item{time}{integer, default = 0}

\item{tempo}{integer, default = 500000}

\item{...}{send additional args to dplyr::add_row}
}
\value{
a new row
}
\description{
Intended for use in a dplyr style pipe. Adds a set_tempo message to a meta header.
}
\examples{
\dontrun{
new_midi_df <- create_empty_midi_df() \%>\%
  add_meta_track_name(name = "bass") \%>\%
  add_meta_tempo(tempo = 500000)
}
}
