% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-construct.R
\name{add_meta_time_sig}
\alias{add_meta_time_sig}
\title{Add time_signature message to df}
\usage{
add_meta_time_sig(
  meta_df,
  i_track = 0,
  type = "time_signature",
  time = 0,
  numerator = 4,
  denominator = 4,
  clocks_per_click = 36,
  notated_32nd_notes_per_beat = 8,
  ...
)
}
\arguments{
\item{meta_df}{midi_df to add}

\item{i_track}{integer, default = 0}

\item{type}{character, default = "time_signature"}

\item{time}{integer, default = 0}

\item{numerator}{integer, default = 4}

\item{denominator}{integer, default = 4}

\item{clocks_per_click}{integer, default = 36}

\item{notated_32nd_notes_per_beat}{integer, default = 8}

\item{...}{send additional args to dplyr::add_row}
}
\value{
a new row
}
\description{
Intended for use in a dplyr style pipe. Adds a time_signature message to a meta header.
}
\examples{
\dontrun{
new_midi_df <- create_empty_midi_df() \%>\%
  add_meta_track_name(name = "bass") \%>\%
  add_meta_tempo(tempo = 500000) \%>\%
  add_meta_time_sig()
}
}
