% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-construct.R
\name{add_meta_track_name}
\alias{add_meta_track_name}
\title{Add track name message to df}
\usage{
add_meta_track_name(
  meta_df,
  i_track = 0,
  type = "track_name",
  name = "",
  time = 0,
  ...
)
}
\arguments{
\item{meta_df}{midi_df to add}

\item{i_track}{integer, default = 0}

\item{type}{character, default = "track_name"}

\item{name}{character, default = ""}

\item{time}{integer, default = 0}

\item{...}{send additional args to dplyr::add_row}
}
\value{
a new row
}
\description{
Intended for use in a dplyr style pipe. Adds a track_name message to a meta header.
}
\examples{
\dontrun{
new_midi_df <- create_empty_midi_df() \%>\%
  add_meta_track_name(name = "bass")
}
}
