% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-construct.R
\name{add_midi_message}
\alias{add_midi_message}
\title{Add a midi message to a midi_df}
\usage{
add_midi_message(
  df,
  i_track = 0,
  meta = FALSE,
  type = "",
  name = NA,
  time = 0,
  numerator = NaN,
  denominator = NaN,
  clocks_per_click = NaN,
  notated_32nd_notes_per_beat = NaN,
  program = NaN,
  channel = NaN,
  control = NaN,
  value = NaN,
  note = NaN,
  velocity = NaN,
  tempo = NaN,
  ...
)
}
\arguments{
\item{df}{midi_df to add row to}

\item{i_track}{integer, default = 0}

\item{meta}{logical, default = FALSE}

\item{type}{character, default = "control_change"}

\item{name}{character, default = NA}

\item{time}{integer, default = 0}

\item{numerator}{integer, default = NaN}

\item{denominator}{integer, default = NaN}

\item{clocks_per_click}{integer, default = NaN}

\item{notated_32nd_notes_per_beat}{integer, default = NaN}

\item{program}{integer, default = NaN}

\item{channel}{integer, default = NaN}

\item{control}{integer, default = NaN}

\item{value}{integer, default = NaN}

\item{note}{integer, default = NaN}

\item{velocity}{integer, default = NaN}

\item{tempo}{integer, default = NaN}

\item{...}{send additional args to dplyr::add_row}
}
\value{
a new row
}
\description{
Intended for use in a dplyr style pipe. Adds a midi message to a midi_df.
}
\examples{
\dontrun{
}
}
