% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-construct.R
\name{add_program_change}
\alias{add_program_change}
\title{Add program_change message to df}
\usage{
add_program_change(
  df,
  i_track = 0,
  type = "program_change",
  time = 0,
  program = 0,
  channel = 0,
  ...
)
}
\arguments{
\item{df}{midi_df to add row to}

\item{i_track}{integer, default = 0}

\item{type}{character, default = "program_change"}

\item{time}{integer, default = 0}

\item{program}{integer, default = 0}

\item{channel}{integer, default = 0}

\item{...}{send additional args to dplyr::add_row}
}
\value{
a new row
}
\description{
Intended for use in a dplyr style pipe. Adds a program_change message to a midi_df.
}
\examples{
\dontrun{
}
}
