% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{bresenham_euclidean}
\alias{bresenham_euclidean}
\title{Bresenham's algorithm for Euclidean rhythm generation}
\usage{
bresenham_euclidean(beats, steps, start = 1)
}
\arguments{
\item{beats}{Integer, number of beats to distribute.}

\item{steps}{Integer, total number of steps across which the beats are to be distributed.}

\item{start}{Integer, the starting value in sequence. Default is 1.}
}
\value{
A numerical vector of size \code{steps} containing 1s where beats are present and 0s elsewhere.
}
\description{
The algorithm evenly distributes a specified number of beats in a total specified number of steps, returning a binary vector where 1s indicate the occurrence of a beat.
}
\details{
Converted to R from discussions of Euclidean Algorithms in this post \url{https://medium.com/code-music-noise/euclidean-rhythms-391d879494df}. See also Toussaint, Godfried. 2005. “The Euclidean Algorithm Generates Traditional Musical Rhythms.” In, 4756. \url{https://archive.bridgesmathart.org/2005/bridges2005-47.html}
}
\examples{
\dontrun{
bresenham_euclidean(5, 16)
bresenham_euclidean(7, 20, 2)
}
}
