% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-construct.R
\name{create_empty_midi_df}
\alias{create_empty_midi_df}
\title{Initialize a midi_df tibble with column headers}
\usage{
create_empty_midi_df()
}
\value{
A tibble object with columns of different types.
}
\description{
This function creates an empty tibble containing specific columns for a midi_df.
}
\details{
The columns include i_track (integer), meta (logical), type (character),
name (character), time (integer), numerator (integer), denominator (integer),
clocks_per_click (integer), notated_32nd_notes_per_beat (integer), program (integer),
channel (integer), control (integer), value (integer), note (integer), velocity (integer),
and tempo (integer).
}
\examples{
\dontrun{
new_midi_df <- create_empty_midi_df()
}

}
