% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{feature_vector_to_matrix}
\alias{feature_vector_to_matrix}
\title{Convert feature vector to matrix}
\usage{
feature_vector_to_matrix(vec, num_notes)
}
\arguments{
\item{vec}{A vector of features. The length of the vector must be divisible by the number of notes.}

\item{num_notes}{An integer specifying the number of notes. This would be the number of rows in the returned matrix.}
}
\value{
A matrix with num_notes rows and length(vec)/num_notes columns.
}
\description{
This function takes in a feature vector and the number of notes, and returns a matrix. Each note is a row of the matrix.
}
