% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{get_feature_probs}
\alias{get_feature_probs}
\title{Get Feature Probabilities}
\usage{
get_feature_probs(midi_matrix)
}
\arguments{
\item{midi_matrix}{A numeric note by time feature matrix.}
}
\value{
A numeric vector containing the probabilities of each feature. The length of the vector is equal to the number of features (columns) in the input MIDI matrix.
}
\description{
This function calculates the probability of each feature as a proportion of total. It sums up all the values in the MIDI matrix, calculates the summary for each column(feature), and finally computes the probability of each feature by dividing the column summary by the total summary.
}
\examples{
\dontrun{
}
}
