% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{get_mean_note_density}
\alias{get_mean_note_density}
\title{Compute the mean note density of a midi matrix}
\usage{
get_mean_note_density(midi_matrix)
}
\arguments{
\item{midi_matrix}{A numerical matrix where the of rows represents the MIDI notes, and the sum of each row represents the note density.}
}
\value{
A rounded integer that represents the mean note density.
}
\description{
This function takes a midi matrix and returns the rounded mean of the sum of rows. Note that the round function is used to round to the nearest whole number.
}
