% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_midi_meta_df}
\alias{get_midi_meta_df}
\title{get meta messages from a track_df}
\usage{
get_midi_meta_df(track_df)
}
\arguments{
\item{track_df}{filtered midi_df with a selected i_track}
}
\value{
tibble of meta messages
}
\description{
get meta messages from a track_df
}
