% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{make_metric_tibble}
\alias{make_metric_tibble}
\title{Create a tibble (dataframe) defining musical metrics}
\usage{
make_metric_tibble(df, ticks_per_beat, bars = NULL, smallest_tick = NULL)
}
\arguments{
\item{df}{A dataframe containing at least a 'time' column and a 'total_time' column. Usually the df from \code{copy_and_extend_midi_df()}}

\item{ticks_per_beat}{Integer specifying the number of ticks per beat}

\item{bars}{Integer specifying the number of bars. If this parameter is NULL, it will compute it assuming 4/4 meter. Default is NULL.}

\item{smallest_tick}{Numeric. This is the smallest unit of time (in beats) represented in the output dataframe. If NULL, it will compute the smallest observed unit in the 'time' column. Default is NULL.}
}
\value{
A dataframe where each row represents one 'tick' (i.e., fundamental temporal unit), and columns represent the 'tick' timestamp (time_steps), its corresponding bar number (bars), and its step within the bar (bar_steps).
}
\description{
This function computes a dataframe that contains the timestamp (in beats, or some other unit), the bar number, and the step within the bar corresponding to each timestamp given the specific parameters of ticks_per_beat (how many ticks (time-resolution steps) per beat are there), the smallest_tick (which determines the resolution of the output), and the total number of bars.
}
\examples{
\dontrun{
copy_track <- copy_and_extend_midi_df(midi_df, track_num = 0)
metric_tibble <- make_metric_tibble(copy_track, ticks_per_beat = 96, bars=48,smallest_tick=8)

}
}
