% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{midi_df_to_matrix}
\alias{midi_df_to_matrix}
\title{Convert MIDI Data Frame to piano roll Matrix}
\usage{
midi_df_to_matrix(midi_df, track = 0)
}
\arguments{
\item{midi_df}{A data frame of MIDI events, from \code{midi_to_object()}}

\item{track}{The track number to be converted to a matrix. Default is 0.}
}
\value{
A binary matrix where column indices represent time frames and row indices represent MIDI note numbers. Each entry in the matrix is 1 if the corresponding note is on at the corresponding time, and 0 otherwise.
}
\description{
This function takes a data frame obtained from a MIDI file and converts it to a binary matrix representation, with time frames as columns and MIDI note numbers (0-127) as rows.
}
\examples{
\dontrun{
}
}
