% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi-construct.R
\name{midi_notes}
\alias{midi_notes}
\title{Create a tibble of midi note information}
\usage{
midi_notes()
}
\value{
tibble with columns for note names, octaves, and midi number
}
\description{
Create a tibble of midi note information
}
\examples{
midi_notes <- midi_notes()
head(midi_notes)
}
