% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{midi_to_object}
\alias{midi_to_object}
\title{midi_to_object Function}
\usage{
midi_to_object(file_path)
}
\arguments{
\item{file_path}{A character string specifying the path to the MIDI file.}
}
\value{
A list of MIDI import objects, which includes:
\itemize{
\item \strong{miditapyr_object}: A miditapyr object representing the complete MIDI file.
\item \strong{mido_object}: A mido object representing the MIDI file.
\item \strong{message_list_df}: A data frame of MIDI messages.
\item \strong{ticks_per_beat}: The number of ticks per beat.
\item \strong{midi_df}: A data frame of MIDI messages, fully unnested.
}
}
\description{
Import MIDI data and transform it into several useful R objects.
}
\details{
This mostly wraps pyramidi functions that call miditapyr and mido.
}
\examples{
\dontrun{
  midi_objects <- midi_to_object("path_to_your_midi_file.midi")
  list2env(midi_objects, .GlobalEnv) # add to global environment
}
}
