% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{reshape_piano_roll}
\alias{reshape_piano_roll}
\title{Reshape Piano Roll Matrix}
\usage{
reshape_piano_roll(piano_roll_matrix, ticks_per_interval)
}
\arguments{
\item{piano_roll_matrix}{A matrix to be reshaped. The matrix rows represent the pitch (from 0 to 127) and the columns represents the time (ticks).}

\item{ticks_per_interval}{A number that specifies number of ticks per interval.}
}
\value{
A reshaped matrix with \code{num_intervals} rows and \code{(ticks_per_interval*128)} columns. Rows represent the time interval and columns represents the pitch.
}
\description{
This function reshapes a piano roll matrix into a specified interval length. It pads with zero or crops the piano roll matrix to fit the desired length.
}
\examples{
\dontrun{
}

}
