% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{set_midi_tempo_meta}
\alias{set_midi_tempo_meta}
\title{set midi tempo in meta messages}
\usage{
set_midi_tempo_meta(meta_messages_df, update_tempo = 5e+05)
}
\arguments{
\item{meta_messages_df}{a df of meta messages}

\item{update_tempo}{Integer, a midi tempo to update}
}
\value{
updated meta_messages_df
}
\description{
set midi tempo in meta messages
}
