% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{set_midi_tempo_midi_df}
\alias{set_midi_tempo_midi_df}
\title{set midi tempo in a midi_df}
\usage{
set_midi_tempo_midi_df(midi_df, update_tempo = 5e+05)
}
\arguments{
\item{midi_df}{a df returned by \code{midi_to_object} or a copy of one}

\item{update_tempo}{Integer, a midi tempo to update}
}
\value{
updated meta_messages_df
}
\description{
set midi tempo in a midi_df
}
