% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{split_meta_df}
\alias{split_meta_df}
\title{Split a meta_df into list of top and end messages}
\usage{
split_meta_df(meta_messages_df)
}
\arguments{
\item{meta_messages_df}{meta_messages_df to split}
}
\value{
list of two tibbles
}
\description{
Split a meta_df into list of top and end messages
}
