% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{copy_midi_df_track}
\alias{copy_midi_df_track}
\title{copy single track df within midi_df}
\usage{
copy_midi_df_track(midi_df, track_num = 0)
}
\arguments{
\item{midi_df}{A dataframe representing a midi file returned from \code{midi_to_object()}.}

\item{track_num}{An integer indicating track number to select. Default is 0.}
}
\value{
A filtered data frame containing messages for the selected track
}
\description{
Copy the midi_df, add cumulative time column, keep only note_on messages
}
\details{
This function filters midi_df for a specific track
}
\examples{
\dontrun{
copy_track_0 = copy_midi_df_track(midi_df, track_num = 0)
}
}
